Add files via upload
My assignment file is Vivianti Santosa - Unit 02 - VBA homework.bas. 
The macro that need to be call to run the calculations is MACRO_TO_RUN

I also attach my Excel file just to show off (a bit), the extra things that I did (creating interface to run the sheet from an opening worksheet, before I realized that my excel file should not be part of the submission. I hope you enjoy it even that it doesn't effect any of the grading.

vv